+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## TODO
 Maybe incorporate aliases into superlineage/sublineage detection. See https://github.com/cov-lineages/pango-designation/blob/master/pango_designation/alias_key.json for more.
 
+## unversioned updates
+### Added
+* plots to summarize lineage detection when abundance values aren't relevent
+* boxplot capability
+
 ## v0.4.0
 ### Added
 * can now plot lineage proportions over vs time